Update stuff.json, test.py and add test2.py
test.py if functional but not performant. the user need to restart the program each time he want to make an another roll. The test2.py use a while loop but there is an issue and the object randomized is the same until the user quit and restart the program. test2.py allows the user to change the target rarety.